created/moved methods aufbauSubjectsDir, aufbauSubjectPath, ensuredirSub, ensuredirSes, ensuredirScans, lns_surfer, lns_tracers

+++ src/+mlnipet/SubjectData.m
@@ -36,6 +36,68 @@ classdef SubjectData < mlpipeline.SubjectData
         
         %%
         
+        function        aufbauSubjectsDir(this)
+            %% e. g., /subjectsDir/{sub-S123456, sub-S123457, ...}
+            
+            S = this.subjectsJson_;
+            for sub = fields(S)'
+                d = this.ensuredirSub(S.(sub{1}).sid);
+                this.aufbauSubjectPath(d, S.(sub{1}));
+            end
+        end
+        function        aufbauSubjectPath(this, sub_pth, S_sub)
+            %% e. g., /subjectsDir/sub-S40037/{ses-E182819, ses-E182853, ...}/tracer.4dfp.*, with sym-linked tracer.4dfp.*
+            
+            if isfield(S_sub, 'aliases')
+                for asub = fields(S_sub.aliases)'
+                    this.aufbauSubjectPath(sub_pth, S_sub.aliases.(asub{1}));
+                end
+            end
+            
+            % base case
+            assert(isfield(S_sub, 'experiments'))
+            for e = S_sub.experiments'
+                d = this.ensuredirSes(sub_pth, e{1});
+                fcell = this.ensuredirsScans(d);
+                if (~isempty(fcell))
+                    e1 = this.experimentID_to_ses(e{1});
+                    try
+                        this.lns_tracers( ...
+                            fullfile(this.projectData.getProjectPath(e1), e1, ''), ...
+                            fullfile(sub_pth, e1, ''), ...
+                            fcell);
+                        this.lns_surfer( ...
+                            fullfile(this.projectData.getProjectPath(e1), e1, ''), ...
+                            fullfile(sub_pth, e1, ''));
+                    catch ME
+                        handwarning(ME);
+                    end
+                end
+            end
+        end
+        function d    = ensuredirSub(this, sid)
+            %  @return d is f.q. path to subject
+
+            d = fullfile(this.subjectsDir, this.subjectID_to_sub(sid), '');
+            ensuredir(d);
+        end
+        function d    = ensuredirSes(this, sub_pth, eid)
+            %  @return d is f.q. path to session
+            
+            d = fullfile(sub_pth, this.experimentID_to_ses(eid), '');
+            ensuredir(d);            
+        end
+        function fcell = ensuredirsScans(this, sub_ses_pth)
+            %  @return fcell is cell of scan-folders
+            
+            p = this.prj_ses_pth_from(sub_ses_pth);
+            prj_ses_scn = cellfun(@(x) fullfile(p, [x '*-Converted']), this.TRACERS, 'UniformOutput', false);
+            dtt = mlpet.DirToolTracer('tracer', prj_ses_scn, 'ac', true);
+            fcell = dtt.dns;
+            for id = fcell
+                ensuredir(fullfile(sub_ses_pth, id{1}));
+            end
+        end
         function ses  = experimentID_to_ses(~, eid)
             split = strsplit(eid, '_');
             ses = ['ses-' split{2}];
@@ -47,6 +109,7 @@ classdef SubjectData < mlpipeline.SubjectData
             pth = this.projectData_.getProjectPath(ses);
         end
         function        lns_surfer(this, prj_ses_pth, sub_ses_pth)
+            %% sym-links project-session surfer objects to subject-session path
             
             % convert /projectsPath/PROJ_00123/ses-E123456/mri/wmparc.mgz
             system(sprintf('mri_convert %s.mgz %s.nii', ...
@@ -59,13 +122,17 @@ classdef SubjectData < mlpipeline.SubjectData
             % ln -s
             for s = {'aparcA2009sAseg' 'aparcAseg' 'brainmask' 'wmparc' 'T1001'}
                 for x = [this.EXTS '.nii']
-                    system(sprintf('ln -s %s%s %s%s', ...
-                        fullfile(prj_ses_pth, s{1}), x{1}, ...
-                        fullfile(sub_ses_pth, s{1}), x{1}));
+                    if ~isfile([fullfile(sub_ses_pth, s{1}) x{1}])
+                        assert(isfile([fullfile(prj_ses_pth, s{1}) x{1}]))
+                        system(sprintf('ln -s %s%s %s%s', ...
+                            fullfile(prj_ses_pth, s{1}), x{1}, ...
+                            fullfile(sub_ses_pth, s{1}), x{1}))
+                    end
                 end
             end
         end
         function        lns_tracers(this, prj_ses_pth, sub_ses_pth, scncell)
+            %% sym-links tracers in project-session path to subject-session path
             %  @param prj_ses_pth is f.q. path
             %  @param sub_ses_pth is f.q. path
             %  @param fcell is cell of scan-folders for prj_ses_pth and sub_ses_pth
@@ -78,20 +145,26 @@ classdef SubjectData < mlpipeline.SubjectData
                     if strncmpi(t{1}, scn{1}, length(t{1}))                        
                         try
                             tracerfp = this.tracer_fileprefix(prj_ses_pth, scn{1}, t{1});
-                            lns_4dfp( ...
-                                fullfile(prj_ses_pth, scn{1}, tracerfp), ...
-                                fullfile(sub_ses_pth, scn{1}, t{1}));
-                            lns_4dfp( ...
-                                fullfile(prj_ses_pth, scn{1}, [tracerfp '_avgt']), ...
-                                fullfile(sub_ses_pth, scn{1}, [t{1} '_avgt']));
+                            if ~lexist_4dfp(fullfile(sub_ses_pth, scn{1}, t{1}))
+                                lns_4dfp( ...
+                                    fullfile(prj_ses_pth, scn{1}, tracerfp), ...
+                                    fullfile(sub_ses_pth, scn{1}, t{1}))
+                            end
+                            if ~lexist_4dfp(fullfile(sub_ses_pth, scn{1}, [t{1} '_avgt']))
+                                lns_4dfp( ...
+                                    fullfile(prj_ses_pth, scn{1}, [tracerfp '_avgt']), ...
+                                    fullfile(sub_ses_pth, scn{1}, [t{1} '_avgt']))
+                            end
                         catch ME
                             handwarning(ME)
                         end                        
                         try
                             t1fp = this.T1001_fileprefix(prj_ses_pth, scn{1}, t{1});
-                            lns_4dfp( ...
-                                fullfile(prj_ses_pth, scn{1}, t1fp), ...
-                                fullfile(sub_ses_pth, scn{1}, 'T1001'));
+                            if ~lexist_4dfp(fullfile(sub_ses_pth, scn{1}, 'T1001'))
+                                lns_4dfp( ...
+                                    fullfile(prj_ses_pth, scn{1}, t1fp), ...
+                                    fullfile(sub_ses_pth, scn{1}, 'T1001'))
+                            end
                         catch ME
                             handwarning(ME)
                         end
@@ -141,29 +214,6 @@ classdef SubjectData < mlpipeline.SubjectData
     end
     
     methods (Access = protected)
-        function d = ensuredirSub(this, sid)
-            %  @return d is f.q. path to subject
-
-            d = fullfile(this.subjectsDir, this.subjectID_to_sub(sid), '');
-            ensuredir(d);
-        end
-        function d = ensuredirSes(this, sub_pth, eid)
-            %  @return d is f.q. path to session
-            
-            d = fullfile(sub_pth, this.experimentID_to_ses(eid), '');
-            ensuredir(d);            
-        end
-        function fcell = ensuredirsScans(this, sub_ses_pth)
-            %  @return fcell is cell of scan-folders
-            
-            p = this.prj_ses_pth_from(sub_ses_pth);
-            prj_ses_scn = cellfun(@(x) fullfile(p, [x '*-Converted']), this.TRACERS, 'UniformOutput', false);
-            dtt = mlpet.DirToolTracer('tracer', prj_ses_scn, 'ac', true);
-            fcell = dtt.dns;
-            for id = fcell
-                ensuredir(fullfile(sub_ses_pth, id{1}));
-            end
-        end
     end
 
 	%  Created with Newcl by John J. Lee after newfcn by Frank Gonzalez-Morphy
